make spacegraphcats.click the default interface
