+++ tests/test_twofoo_short.py
@@ -11,7 +11,7 @@ import sourmash
 import screed
 
 from spacegraphcats.search.catlas import CAtlas
-from spacegraphcats.click import run_snakemake
+from spacegraphcats.__main__ import run_snakemake
 from . import pytest_utils as utils
 
 # NOTE re dependencies (@pytest.mark.dependency):
